# api/openapi.yaml
openapi: "3.0.3"
info:
  title: "InferGlow API"
  description: "InferGlow Agent Framework REST API"
  version: "6.0.0"
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /v1/memories/{id}:
    get:
      operationId: getMemory
      summary: Get a memory by ID
      tags: [Memories]
      parameters:
        - $ref: "#/components/parameters/MemoryId"
      responses:
        "200":
          description: Memory record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryRecord"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      operationId: deleteMemory
      summary: Delete a memory by ID
      tags: [Memories]
      parameters:
        - $ref: "#/components/parameters/MemoryId"
      responses:
        "200":
          description: Memory deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryDeleteResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  parameters:
...
    MemoryId:
      name: id
      in: path
      required: true
      description: Memory ID
      schema:
        type: string
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServiceUnavailable:
      description: Service not configured
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required: [error]
...
    MemoryRecord:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        category:
          type: string
        facts:
          type: array
          items:
            type: string
...
    MemoryDeleteResponse:
      type: object
      properties:
        status:
          type: string
          example: "deleted"
        id:
          type: string
...